# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Book Platform API - Complete Reading & Recommendation System
  description: |
    Production-ready API for a comprehensive book platform featuring:
    - Book catalog management
    - PDF content processing with AI
    - Reading progress tracking
    - AI-powered chat assistant
    - Personalized recommendations
    - User interactions & ratings
  version: 2.0.0
  contact:
    name: Leonard Oseghale
    email: your-email@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.yourplatform.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/register
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          example: "john@example.com"
        profileImage:
          type: string
          example: "https://cloudinary.com/..."
        preferredGenres:
          type: array
          items:
            type: string
          example: ["fiction", "mystery"]
        favoriteAuthors:
          type: array
          items:
            type: string
          example: ["Agatha Christie", "Stephen King"]
    Book:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        title:
          type: string
          example: "The Great Gatsby"
        subTitle:
          type: string
          example: "A Novel"
        author:
          type: string
          example: "F. Scott Fitzgerald"
        caption:
          type: string
          example: "A classic American novel"
        description:
          type: string
          example: "The story primarily concerns the young..."
        genres:
          type: array
          items:
            type: string
          example: ["fiction", "classic", "romance"]
        image:
          type: string
          example: "https://cloudinary.com/..."
        price:
          type: number
          example: 9.99
        isbn:
          type: string
          example: "978-0-7432-7356-5"
        publishedYear:
          type: number
          example: 1925
        averageRating:
          type: number
          example: 4.5
        totalRatings:
          type: number
          example: 1234
        totalViews:
          type: number
          example: 5678
        totalPurchases:
          type: number
          example: 890
        hasContent:
          type: boolean
          example: true
        status:
          type: string
          enum: [draft, published, archived]
          example: "published"
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '401':
          description: Invalid credentials

  # ==================== BOOKS - CATALOG ====================
...
  /recommendations/personalized:
    get:
      summary: Get personalized recommendations
      tags: [Recommendations]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
...
  /recommendations/popular:
    get:
      summary: Get popular books
      tags: [Recommendations]
      security: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Popular books
  /recommendations/new:
    get:
      summary: Get new releases
      tags: [Recommendations]
      security: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: New books
  /recommendations/trending:
    get:
      summary: Get trending books
      tags: [Recommendations]
      security: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: days
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Trending books
  /recommendations/genre/{genre}:
    get:
      summary: Get books by genre
      tags: [Recommendations]
      security: []
      parameters:
        - in: path
          name: genre
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Books in genre
  /recommendations/author/{author}:
    get:
      summary: Get books by author
      tags: [Recommendations]
      security: []
      parameters:
        - in: path
          name: author
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Books by author